+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 > Features
 - **filetype:** env file is now considered sh
 - **gps:** added augroup
+- **null-ls:** shellcheck runtime not run on .env files
 
 > Revert
 - winbar only appears when gps exists